requirements.txt
pdfplumber
openai>=1.0.0
requests
streamlit
python-dotenv 